+++ src/Components/drawer/blogs/blogsList/index.jsx
@@ -8,7 +8,7 @@ import useSessionExpired from '../../../../customHooks/useSessionExpired';
 import useActivePageState from '../../../../customHooks/useAcitvePageState';
 import LoadingCardArray from '../../../common/LoadingCardArray';
 import BlogsArray from '../../../user/myBlogs/BlogsArray';
-import { allBlogsLimit } from '../../../../constants';
+import { allBlogsLimit } from '../../../../utils/constants';
 
 const AllBlogsPageContainer = ({ isSuperuserRoute = false }) => {
   const { redirectOnSessionExpiredBeforeRender, isSessionExpired } = useSessionExpired();
