+++ src/utils/commonFunctions.js
